BRANCH POINT
This is the point where I split the Hackathon website from the future DASH tutorial.

The website on main Hackathon branch will be pared down to focus on some demos. This is necessary because of the time constraints of the workshop.

The new dash-tutorial branch should be adapted later into the DASH tutorial within the DASH docs.